Separate command callback and middleware functionality
